extract basic chart builder
