+++ air-quality-ui/src/components/single-city/site-measurement-chart/SiteMeasurementsChart.tsx
@@ -1,11 +1,14 @@
-import { EChartsOption, SeriesOption } from 'echarts'
 import ReactECharts from 'echarts-for-react'
 import { useCallback, useMemo } from 'react'
 
+import {
+  GenerateMeasurementChart,
+  createForecastSeries,
+  createMeasurementSeries,
+} from './site-measurement-chart-builder'
 import classes from './SiteMeasurementsChart.module.css'
 import { useForecastContext } from '../../../context'
 import { PollutantType, pollutantTypeDisplay } from '../../../models'
-import { convertToLocalTime } from '../../../services/echarts-service'
 import { getInSituPercentage } from '../../../services/forecast-time-service'
 import {
   ForecastResponseDto,
@@ -18,87 +21,6 @@ interface SiteMeasurementsChartProps {
   onSiteClick: (siteName: string) => void
   pollutantType: PollutantType
   seriesColorsBySite?: Record<string, string>
-}
-
-const createForecastSeries = (
-  pollutantType: PollutantType,
-  forecastData: ForecastResponseDto[],
-): SeriesOption => {
-  return {
-    type: 'line',
-    data: forecastData.map((data) => [
-      convertToLocalTime(data.valid_time),
-      data[pollutantType].value.toFixed(1),
-    ]),
-    name: 'Forecast',
-    color: 'black',
-    lineStyle: {
-      width: 5,
-      type: 'dashed',
-    },
-    zlevel: 2,
-    z: 2,
-  }
-}
-
-const createMeasurementSeries = (
-  pollutantType: PollutantType,
-  measurementsBySite: Record<string, MeasurementsResponseDto[]>,
-  seriesColorsBySite?: Record<string, string>,
-): SeriesOption[] =>
-  Object.entries(measurementsBySite).map(([siteName, measurements]) => ({
-    type: 'line',
-    data: measurements
-      .filter((f) => f[pollutantType] !== undefined)
-      .map((f) => [
-        convertToLocalTime(f.measurement_date),
-        f[pollutantType]?.toFixed(1),
-      ]),
-    name: siteName,
-    ...(seriesColorsBySite && {
-      color: seriesColorsBySite[siteName],
-    }),
-    triggerLineEvent: true,
-    lineStyle: {
-      width: 1,
-      type: 'solid',
-      opacity: 0.5,
-    },
-    zlevel: 1,
-    z: 1,
-    symbol: 'roundRect',
-  }))
-
-const getChartOptions = (
-  zoomPercent: number,
-  ...series: SeriesOption[]
-): EChartsOption => {
-  return {
-    xAxis: {
-      type: 'time',
-    },
-    yAxis: {
-      type: 'value',
-      name: 'µg/m³',
-      nameGap: 50,
-      nameLocation: 'middle',
-    },
-    series,
-    dataZoom: [
-      {
-        show: true,
-        realtime: false,
-        end: zoomPercent,
-      },
-      {
-        type: 'inside',
-        realtime: false,
-      },
-    ],
-    tooltip: {
-      trigger: 'item',
-    },
-  }
 }
 
 export const SiteMeasurementsChart = ({
@@ -156,16 +78,14 @@ export const SiteMeasurementsChart = ({
 
   return (
     <>
-      <label className={classes['chart-label']}>
-        {pollutantTypeDisplay[pollutantType]}
-      </label>
       <ReactECharts
         className={classes['chart']}
         onEvents={{
           click: eChartEventHandler,
         }}
         notMerge
-        option={getChartOptions(
+        option={GenerateMeasurementChart(
+          pollutantTypeDisplay[pollutantType],
           zoomPercent,
           forecastSeries,
           ...measurementSeries,
